requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2

# Deep Learning (Image CNN)
tensorflow==2.13.0
keras==2.13.1

# Computer Vision
opencv-python==4.8.0.76
Pillow==10.0.0

# Backend
flask==2.3.3
flask-cors==4.0.0

# Frontend
streamlit==1.28.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.3

# Audio
librosa==0.10.1
soundfile==0.12.1

# Utilities
requests==2.31.0
werkzeug==2.3.7